backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.1.3
pdfplumber==0.10.3
camelot-py==0.11.0
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
xlsxwriter==3.1.9
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
ocrmypdf==15.4.4

